streamlit_app/requirements.txt
streamlit>=1.32.0
ultralytics>=8.4.0
opencv-python-headless>=4.6.0
Pillow>=9.0.0
numpy>=1.23.0
